+++ apps/directorate-of-equality-web/src/components/list-page/Filter/CaseFilter.css.ts
@@ -1,5 +1,6 @@
 import { style } from '@vanilla-extract/css'
 
+// Pulls the date section up to overlap the Filter component's bottom border
 export const dateSection = style({
   marginTop: '-32px',
   borderBottomLeftRadius: '8px',
