+++ apps/directorate-of-equality-web/src/components/list-page/Tabs/TabContent.css.ts
@@ -1,0 +1,15 @@
+import { style } from '@vanilla-extract/css'
+
+export const expandedRowGrid = style({
+  display: 'flex',
+  flexWrap: 'wrap',
+  columnGap: 16,
+})
+
+export const expandedRowItem = style({
+  flex: '0 0 calc(50% - 8px)',
+})
+
+export const expandedRowLabel = style({
+  minWidth: 220,
+})
